Refactor Searchbar and Clients components, so that Clients.tsx handles the filtering instead of Seachbar.
